requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# Real-time (Socket.IO)
python-socketio[asyncio]>=5.12.0

# Google ADK (includes google-genai as dependency)
google-adk>=1.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.0

# DRL + ML
torch>=2.5.0
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.6.0

# Data processing + file parsing
openpyxl>=3.1.0
pdfplumber>=0.11.0

# Tokenization (for chunking before LLM + vector DB)
tiktoken>=0.12.0

# Task queue (celery[redis] includes redis package)
celery[redis]>=5.4.0

# Auth (full backend auth: JWT + password hashing)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Utilities (pydantic comes with FastAPI, python-dotenv comes with pydantic-settings)
pydantic-settings>=2.7.0
